Use stable Node version for linting on Travis CI
eslint 3.x dropped support for Node < 4, so tests were failing on those
versions. This change switches the Node version used in Travis CI builds to
the current stable version before running eslint, then switches it back to the
version Travis was about to use to run the (not-yet-existent) tests.

This is a local workflow change in that linting is no longer performed when
running 'npm test' and will require running 'npm run lint' manually.