Use npm to run lint script on Travis CI
Since we're using a different version of Node for linting (due to eslint
dropping support for Node < 4), the yarn package we installed prior to the
temporary version switch is not available. It's silly to install it again just
to run a script, so we'll just stick with npm on this step.