build(target): remove support for Node.js v6
BREAKING CHANGE: This removes support for Node v6 and makes the new minimum runtime Node v8, as v6
is EOL as of April, 2019. Please upgrade your Node.js environment to at least version 8, or
continue using the latest v6 release of `pwned` if you are unable to upgrade your environment.