feat(api-v3): invert truncate option default value
This changes the default value of the `truncate` option from `false` to `true` for both the `ba`
and `search` commands.

BREAKING CHANGE: The `truncate` option of the `ba` and `search` commands now defaults to `true`,
per Troy's recommendation. If you do not negate it explicitly, each `Breach` result will only
contain the breach name (no metadata).