chore(node): drop Node 10 support
This also converts the project to pure ESM as it is fully supported in
Node 12.20 and beyond.

BREAKING CHANGE: This removes support for Node v10 and makes the new
minimum runtime Node v12.20, as v10 is EOL. Please upgrade your Node.js
environment to at least version 12.20, or continue using the latest v9
release of `pwned` if you are unable to upgrade your environment.