chore(node): drop Node 12 support
This also converts the project to pure ESM as it is fully supported in
Node 14.13.1, 16 and beyond.

BREAKING CHANGE: This removes support for Node v12 and makes the new
minimum runtime Node v14.13.1, as v12 is EOL. Please upgrade your
Node.js environment to at least version 14.13.1, or continue using the
latest v9 release of `pwned` if you are unable to upgrade your
environment.